Added page8kpis tests
